+++ pancake/nbs.h
@@ -191,7 +191,7 @@ class Nbs {
       }
       std::cout << "\n";
 #endif 
-      return std::make_tuple(UB, expansions, memory);
+      return std::make_tuple((double)UB, expansions, memory);
     }
     else return std::make_tuple(std::numeric_limits<double>::infinity(), expansions, memory);
   }
